:hammer: feat: Add endpoint to find not-friends users
Adds a new endpoint /users/not-friends that allows a user to get a paginated list of other users who are not yet their friends.

This includes:

A new getNotFriendsPaginated method in CrudUserService with a MongoDB aggregation to efficiently filter users.

The corresponding UserDto for the response.

The controller endpoint in UserController.